08-power-quality-monitor/dashboard/requirements.txt
altair
pandas
streamlit==1.23.0
snowflake-connector-python==2.7.6
matplotlib